Handle legacy clusters restore
